# docs/swagger.yaml
basePath: /
definitions:
  bookings.CreateBookingResponse:
    properties:
      id:
        type: integer
      total_payment:
        type: integer
      total_ticket:
        type: integer
      valid_until:
        type: string
    type: object
  events.EventDetail:
    properties:
      created_at:
        type: string
      event_time:
        type: string
      id:
        type: integer
      name:
        type: string
      ticket_booked:
        type: integer
      ticket_price:
        type: string
      ticket_remains:
        type: integer
      total_tickets:
        type: integer
    type: object
  events.EventList:
    properties:
      created_at:
        type: string
      event_time:
        type: string
      id:
        type: integer
      name:
        type: string
      ticket_price:
        type: string
      total_tickets:
        type: integer
    type: object
  requests.CreateBooking:
    properties:
      event_id:
        type: integer
      name:
        type: string
      total:
        type: integer
    required:
    - event_id
    - name
    - total
    type: object
  requests.CreateEvent:
    properties:
      event_time:
        type: string
      name:
        type: string
      ticket_price:
        type: integer
      total_tickets:
        type: integer
    required:
    - event_time
    - name
    - ticket_price
    - total_tickets
    type: object
  responses.EventList:
    properties:
      events:
        items:
          $ref: '#/definitions/events.EventList'
        type: array
    type: object
  responses.MyError:
    properties:
      message:
        type: string
    type: object
  responses.ResponseId:
    properties:
      id:
        type: integer
    type: object
host: localhost:11300
info:
  contact: {}
  title: Ticket booking API documentation
  version: 0.1.0
paths:
  /events:
    get:
      description: Get all events
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.EventList'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.MyError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.MyError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.MyError'
      summary: Get all events
      tags:
      - Events
    post:
      description: Create event
      parameters:
      - description: Event Data
        in: body
        name: event
        required: true
        schema:
          $ref: '#/definitions/requests.CreateEvent'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.ResponseId'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.MyError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.MyError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.MyError'
      summary: Create event
      tags:
      - Events
  /events/{id}:
    get:
      description: Get event detail
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/events.EventDetail'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.MyError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.MyError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.MyError'
      summary: Get event detail
      tags:
      - Events
  /tickets/booking:
    post:
      description: Booking tickets for an event
      parameters:
      - description: Booking Data
        in: body
        name: event
        required: true
        schema:
          $ref: '#/definitions/requests.CreateBooking'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/bookings.CreateBookingResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.MyError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.MyError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.MyError'
      summary: Booking tickets for an event
      tags:
      - Tickets
swagger: "2.0"